Updated tab key tests (moved to TableNavigation plugin).
